refactor fast conv
